// src/app/pages/maliyet/recete/models/recete-create-model.ts
export class ReceteCreateModel {
  kod?: any;
  ad?: any;
  adet?: any;
  miktar?: any;
  olcuAciklama?: any;
  receteTuru?: any;
  etkinlik?: any;
  durum?: any;
  active?: any;
  stokId?: any;
  parentId?: any;

}


// public string Ad { get; set; }
// public string Kod { get; set; }
// public decimal Adet { get; set; }
// public decimal Miktar { get; set; }
// public string OlcuAciklama { get; set; }
// public string ReceteTuru { get; set; }
// public decimal Etkinlik { get; set; }
// public bool Durum { get; set; }
// public bool Active { get; set; }
// public Guid? StokId { get; set; }
// public Guid? ParentId { get; set; }